+++ packages/expo-router/src/native-tabs/common/elements.tsx
@@ -164,3 +164,32 @@ export interface NativeTabsTriggerBadgeProps {
 }
 
 export const NativeTabsTriggerBadge: React.FC<NativeTabsTriggerBadgeProps> = Badge;
+
+export interface NativeTabsBottomAccessoryProps {
+  children?: React.ReactNode;
+}
+
+/**
+ * A [bottom accessory](https://developer.apple.com/documentation/uikit/uitabbarcontroller/bottomaccessory) for `NativeTabs` on iOS 26 and above.
+ *
+ * @example
+ * ```tsx
+ * import { NativeTabs } from 'expo-router/unstable-native-tabs';
+ *
+ * export default Layout(){
+ *   return (
+ *     <NativeTabs>
+ *       <NativeTabs.BottomAccessory>
+ *         <YourAccessoryComponent />
+ *       </NativeTabs.BottomAccessory>
+ *       <NativeTabs.Trigger name="index" />
+ *     </NativeTabs>
+ *   );
+ * }
+ * ```
+ *
+ * @platform iOS 26+
+ */
+export const NativeTabsBottomAccessory: React.FC<NativeTabsBottomAccessoryProps> = () => {
+  return null;
+};